requirements.txt
Scrapy==1.8.0
apyori==1.1.2
pandas==0.25.3
beautifulsoup4==4.8.2
config==0.4.2
google_api_python_client==1.7.11
pyenchant==2.0.0
PyPDF2==1.26.0
scikit_learn==0.22.1